Upload car.erl
I demonstrate object oriented programming in Erlang with my car blueprint that I use in some of my other projects. Erlang is very immutable, and so new data items need to be created instead of updating the existing data item. The exception is in extending the data item, but that is not relevant for this example. In this example, records are used in place of classes or structs, and member functions are not supported.